Blackjack rework Reworked some special rooms Fixed patch room levelgen bug

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/ArmoredStatue.java
@@ -24,6 +24,7 @@ package com.shatteredpixel.shatteredpixeldungeon.actors.mobs;
 import com.shatteredpixel.shatteredpixeldungeon.Challenges;
 import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
 import com.shatteredpixel.shatteredpixeldungeon.actors.Char;
+import com.shatteredpixel.shatteredpixeldungeon.actors.buffs.NoReward;
 import com.shatteredpixel.shatteredpixeldungeon.items.Generator;
 import com.shatteredpixel.shatteredpixeldungeon.items.armor.Armor;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
@@ -100,12 +101,14 @@ public class ArmoredStatue extends Statue {
 
 	@Override
 	public void die( Object cause ) {
-		armor.identify();
-		if(Challenges.CURSED.enabled()){
-			armor.cursed=true;
-			armor.inscribe(Armor.Glyph.randomCurse());
+		if (buff(NoReward.class) == null) {
+			armor.identify();
+			if (Challenges.CURSED.enabled()) {
+				armor.cursed = true;
+				armor.inscribe(Armor.Glyph.randomCurse());
+			}
+			Dungeon.level.drop(armor, pos).sprite.drop();
 		}
-		Dungeon.level.drop( armor, pos ).sprite.drop();
 		super.die( cause );
 	}
